refactor: rename 'Horario' classes to 'Hora'
